% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_first_word.R
\name{abbreviate_first_word}
\alias{abbreviate_first_word}
\title{Abbreviate first word of string}
\usage{
abbreviate_first_word(string)
}
\description{
Transforms e.g. "Akkermansia muciniphila" to "A. muciniphila"
Doesn't transform string if only composed by one word
Could be applied in a vectorised form
}
\examples{
\dontrun{
string <- "Akkermansia muciniphila"
abbreviate_first_word(string)
vector <- c("Akkermansia muciniphila", "Bifidobacterium longum subsp. infantis")
as.vector(sapply(vector, abbreviate_first_word, simplify=T))
}
}
