% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suffix_to_duplicates.R
\name{add_suffix_to_duplicates}
\alias{add_suffix_to_duplicates}
\title{Add suffixes to duplicated strings in a vector
If all unique elements in the vector have less than 26 duplicates,
then suffix will be a latin letter, otherwise it will be a roman number}
\usage{
add_suffix_to_duplicates(vector)
}
\description{
Add suffixes to duplicated strings in a vector
If all unique elements in the vector have less than 26 duplicates,
then suffix will be a latin letter, otherwise it will be a roman number
}
\examples{
\dontrun{
vector <- c("hello", "hi", "hola", "hello", "hello")
add_suffix_to_duplicates(vector)
}
}
