% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zero_to_single_digits.R
\name{add_zero_to_single_digits}
\alias{add_zero_to_single_digits}
\title{Add zeroes to the left of some elements (strings) from a vector}
\usage{
add_zero_to_single_digits(vector)
}
\description{
Works for sorting numeric vectors in a string format
Adds as many zeroes as the maximum number of characters observed in all elements of the vector
If max digits in vector elements is 2 (e.g. "89"), and element is "1", it will be transformed to "01"
If max digits in vector elements is 2 (e.g. "89"), and element is "11", it will not be transformed
}
\examples{
\dontrun{
vector <- c("1", "2", "12", "111")
sort(vector)
transformed_vector <- add_zero_to_single_digits(vector)
sort(transformed_vector)
}
}
