% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_first_letter.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize first letter of string}
\usage{
capitalize_first_letter(string)
}
\description{
Capitalize first letter of string
}
\examples{
\dontrun{
capitalize_first_letter("hello world!")
}
}
