% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_na.R
\name{cbind_na}
\alias{cbind_na}
\title{Bind two lists in two columns of a df
Bind two lists in two columns of a df - add NAs when shortest column is fully added}
\usage{
cbind_na(a, b)
}
\value{
a df with 2 columns
}
\description{
Bind two lists in two columns of a df
Bind two lists in two columns of a df - add NAs when shortest column is fully added
}
\examples{
\dontrun{
cbind_na(list1, list2)
cbind_na(vector1, vector2)
}

}
