% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_rows.R
\name{common_rows}
\alias{common_rows}
\title{Common or different items
Show common or different items i.e. of a list, colnames of a df, rows in a df}
\usage{
common_rows(data_1, data_2, colname_1, colname_2 = NULL)
}
\description{
Common or different items
Show common or different items i.e. of a list, colnames of a df, rows in a df
}
\examples{
\dontrun{
common_colnames(df1, df2)
different_colnames(df1, df2)
common_items(list1, list2)
different_items(list1, list2)
common_rows(df1, df2, colname1, colname2) # colname2 can be NULL, then same name as colname1 is taken
}

}
