% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_how_many_above_threshold.R
\name{count_how_many_above_threshold}
\alias{count_how_many_above_threshold}
\title{Count how many values in a vector are above threshold}
\usage{
count_how_many_above_threshold(vector, threshold)
}
\description{
Count how many values in a vector are above threshold
}
\examples{
\dontrun{
vector <- c(1, 2, -1, 2, 3)
count_how_many_above_threshold(vector, 2)
}
}
