% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_how_many_na_per_column.R
\name{count_how_many_na_per_column}
\alias{count_how_many_na_per_column}
\title{NA handy functions - count NAs per column
Counts how many NAs per column}
\usage{
count_how_many_na_per_column(data)
}
\value{
a number
}
\description{
NA handy functions - count NAs per column
Counts how many NAs per column
}
\examples{
\dontrun{
data \%>\% count_how_many_na_per_column()
}

}
