% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_how_many_string_per_column.R
\name{count_how_many_string_per_column}
\alias{count_how_many_string_per_column}
\title{Count how many times a string is detected per column
Counts how many times a string is detected per column in a df}
\usage{
count_how_many_string_per_column(data, string)
}
\value{
a number per column
}
\description{
Count how many times a string is detected per column
Counts how many times a string is detected per column in a df
}
\examples{
\dontrun{
data \%>\% count_how_many_string_per_column("mystring")
}


}
