% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_how_many_zero_per_column.R
\name{count_how_many_zero_per_column}
\alias{count_how_many_zero_per_column}
\title{Zero handy functions - count zeroes per column
Counts how many zeroes per column}
\usage{
count_how_many_zero_per_column(data)
}
\value{
a number
}
\description{
Zero handy functions - count zeroes per column
Counts how many zeroes per column
}
\examples{
\dontrun{
data \%>\% count_how_many_zero_per_column()
}

}
