% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rows_with_na.R
\name{count_rows_with_na}
\alias{count_rows_with_na}
\title{NA handy functions - count rows with NA
Counts number of rows with NAs}
\usage{
count_rows_with_na(data)
}
\value{
a number
}
\description{
NA handy functions - count rows with NA
Counts number of rows with NAs
}
\examples{
\dontrun{
data \%>\% count_rows_with_na()
}

}
