% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different_items.R
\name{different_items}
\alias{different_items}
\title{Common or different items
Show common or different items i.e. of a list, colnames of a df, rows in a df}
\usage{
different_items(list_1, list_2)
}
\description{
Common or different items
Show common or different items i.e. of a list, colnames of a df, rows in a df
}
\examples{
\dontrun{
common_colnames(df1, df2)
different_colnames(df1, df2)
common_items(list1, list2)
different_items(list1, list2)
common_rows(df1, df2, colname1, colname2) # colname2 can be NULL, then same name as colname1 is taken
}

}
