% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{Expand grid with 2 vectors, but only outputs unique combinations
Order of factors is not considered i.e. 1&4 is the same as 4&1 and therefore only one combination will be in the output}
\usage{
expand_grid_unique(x, y, colnames = NULL, include.equals = FALSE)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{colnames}{vector of 2 names for the column names of the resulting dataframe}
}
\value{
dataframe containing the unique combinations of x and y (order of factors is not considered, see above)
}
\description{
Expand grid with 2 vectors, but only outputs unique combinations
Order of factors is not considered i.e. 1&4 is the same as 4&1 and therefore only one combination will be in the output
}
\examples{
\dontrun{
expand_grid_unique(c(1,2,3,1), c(4,5,6))
tibble::as_tibble(expand_grid_unique(c(1,2,3,1), c(4,5,6), c("col1", "col2")))
}
}
