% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_columns_with_all_na.R
\name{get_columns_with_all_na}
\alias{get_columns_with_all_na}
\title{NA handy functions - get columns with all NAs
Shows columns which have all NAs in a tibble}
\usage{
get_columns_with_all_na(data)
}
\value{
a tibble
}
\description{
NA handy functions - get columns with all NAs
Shows columns which have all NAs in a tibble
}
\examples{
\dontrun{
data \%>\% get_columns_with_all_na()
}

}
