% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_columns_with_any_na.R
\name{get_columns_with_any_na}
\alias{get_columns_with_any_na}
\title{NA handy functions - get columns with any NAs
Shows columns which have any NAs in a tibble}
\usage{
get_columns_with_any_na(data)
}
\value{
a tibble
}
\description{
NA handy functions - get columns with any NAs
Shows columns which have any NAs in a tibble
}
\examples{
\dontrun{
data \%>\% get_columns_with_any_na()
}

}
