% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Obtain quantiles of a vector}
\usage{
get_quantiles(vector, breaks = 4, labels = FALSE)
}
\arguments{
\item{vector}{vector}

\item{breaks}{how many quantiles, default 4 (i.e quartiles)}

\item{labels}{either FALSE or NULL, default FALSE}
}
\description{
Obtain quantiles of a vector
}
\examples{
\dontrun{
vector <- c(1, 2, -1, 2, 3)
get_quantiles(vector) # quartiles
get_quantiles(vector, breaks = 3) # tertiles
}
}
