% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_and_filter_vector.R
\name{grep_and_filter_vector}
\alias{grep_and_filter_vector}
\title{Filter a vector with elements matching a pattern}
\usage{
grep_and_filter_vector(vector, pattern)
}
\arguments{
\item{vector}{the vector}

\item{pattern}{the pattern}
}
\description{
Filter a vector with elements matching a pattern
}
\examples{
\dontrun{
vector <- list("hello", "hi", "hola", "hey")
grep_and_filter_vector(vector, "he")
}
}
