% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_analysis.R
\name{ia}
\alias{ia}
\title{Initialize analysis/study}
\usage{
ia()
}
\value{
calls devtools load_all, install, and
cm.initialize_study(author = "Test Demo", demo = TRUE, method = "acmemap4")
and it knits the 010.Rmd
}
\description{
Initialize analysis/study
}
\examples{
ia()
}
