% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc.R
\name{perc}
\alias{perc}
\title{Calculates percentage of first element relative to that element and the remaining elements}
\usage{
perc(n1, n2)
}
\arguments{
\item{n1}{number from which the percentage will be calculated}

\item{n2}{number or vector of numbers, used as demonimator (together with n1) to calculate the percentage of n1}
}
\description{
Calculates percentage of first element relative to that element and the remaining elements
}
\examples{
\dontrun{
perc(12, 24)
perc(12, c(12, 24))
}
}
