% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_to_file_path.R
\name{setwd_to_file_path}
\alias{setwd_to_file_path}
\title{Set working directory to the path of the current script}
\usage{
setwd_to_file_path()
}
\description{
Set working directory to the path of the current script
}
\examples{
\dontrun{
setwd_to_file_path()
}

}
