% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rmd_params.R
\name{show_rmd_params}
\alias{show_rmd_params}
\title{Show R markdown parameters}
\usage{
show_rmd_params(params)
}
\arguments{
\item{params}{list of Rmd parameters}
}
\value{
table showing all the R markdown parameters and their correspinding values
}
\description{
Show R markdown parameters
}
\examples{
\dontrun{
show_rmd_params()
}

}
