% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rows_with_na.R
\name{show_rows_with_na}
\alias{show_rows_with_na}
\title{NA handy functions - show rows with NAs
Shows rows with NAs in a tibble}
\usage{
show_rows_with_na(data)
}
\value{
a tibble
}
\description{
NA handy functions - show rows with NAs
Shows rows with NAs in a tibble
}
\examples{
\dontrun{
data \%>\% show_rows_with_na()
}

}
