% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_signif.R
\name{trunc_signif}
\alias{trunc_signif}
\title{Truncate to a number of significant digits}
\usage{
trunc_signif(x, digits)
}
\value{
number, truncated to the selected number of significant digits
}
\description{
for digits, I recommend using "trunc(log10(min(your_data_column))) + 1"
}
\examples{
\dontrun{
trunc_signif(498, 2) # returns 400
trunc_signif(498, 1) # returns 490
trunc_signif(-498, 2) # returns -500
limits$Shannon  <- c(trunc_signif(min(alphaDiv$Shannon), digits = trunc(log10(min(alphaDiv$Shannon))) + 1), ceiling_signif(max(alphaDiv$Shannon), digits = trunc(log10(max(alphaDiv$Shannon)))))
}

}
