% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_items_are_duplicated.R
\name{which_items_are_duplicated}
\alias{which_items_are_duplicated}
\title{Extract which items in a column of tibble are duplicated}
\usage{
which_items_are_duplicated(tibble, column)
}
\arguments{
\item{tibble}{a tibble}

\item{column}{string, column name of the tibble (must be specified between quotes)}
}
\description{
Extract which items in a column of tibble are duplicated
}
\examples{
\dontrun{
tibble <- tibble::tibble('name' = c("a", "b", "c", "a", "aa", "a", "c"))
which_items_are_duplicated(tibble, "name")
}
}
