% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pvalue_string.R
\name{write_pvalue_string}
\alias{write_pvalue_string}
\title{Write a string showing P-value in a nice format}
\usage{
write_pvalue_string(p, cutoff = 0.01, decimal_round = 2, name = "p")
}
\description{
Write a string showing P-value in a nice format
}
\examples{
\dontrun{
write_pvalue_string(0.00000001)
write_pvalue_string(0.00000001, cutoff = 0.001, decimal_round = 3)
}
}
