% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_theme.R
\name{white_theme}
\alias{white_theme}
\title{Custom ggplot theme with white background}
\usage{
white_theme(
  font_size = 10,
  line_size = 0.5,
  smaller_fraction = 0.8,
  axis_margin_fraction = 0.25
)
}
\arguments{
\item{font_size}{Numeric. Default font size.}

\item{line_size}{Numeric. Default line size.}

\item{smaller_fraction}{Numeric. Some texts (e.g. axes ticks) will be the given fraction of the font size smaller.}

\item{axis_margin_fraction}{Numeric. Axis margin size will be calculated based on this fraction, the "smaller_fraction" and the font size.}
}
\value{
plot (ggplot) with the theme
}
\description{
Custom ggplot theme with white background
}
\examples{
\dontrun{
data = data.frame(x = seq(1,200,10), y = seq(1,400,20))
ggplot2::ggplot(data, ggplot2::aes(x = x, y = y)) + ggplot2::geom_point() + white_theme()
}

}
